Add table of contents

+++ app.js
@@ -1,36 +1,43 @@
 import fs from 'fs'
+import path from 'path'
 import inquirer from 'inquirer'
 import licenses  from './src/licenses.js'
 import { questions, license } from './src/questions.js'
 
+const buildTableOfContents = answers => {
+  let tableOfContents = '## Table of Contents\n'
+  Object.keys(answers).forEach(answer => {
+    tableOfContents += `- [${answer}](#${answer})\n`
+  })
+  return `${tableOfContents}`
+}
+
 const buildReadMe = answers => {
+  // TODO: Create a table of contents!
+  let section = `# ${answers['title']}\n
+![License](https://img.shields.io/badge/License-${license}-white?labelColor=green&style=flat)\n
+${buildTableOfContents(answers)}\n`
   for (const answer in answers) {
-    let section = ''
-    if (['github', 'email', 'questions'].includes(answer)) return
-
-    if (answer === 'title') {
-      section += `# ${answers[answer]}\n`
-    } else {
+    if (!['title', 'github', 'email'].includes(answer)) {
       section += `## ${answer}\n`
+      
+      if (answer === 'License') {
+        section += `${licenses[license]}\n`
+      } else {
+        section += `${answers[answer]}\n`
+      }
+      fs.appendFileSync('./README.md', section, err => {
+        if (err) console.error(err)
+      })
+      section = ''
     }
-
-    if (answer === 'title') {
-      section += `![License](https://img.shields.io/badge/License-${license}-white?labelColor=green&style=flat)\n`
-    } else if (answer === 'License') {
-      section += `${licenses[license]}\n`
-    } else {
-      section += `${answers[answer]}\n`
-    }
-    fs.appendFileSync('./README.md', section, err => {
-      if (err) console.error(err)
-    })
   }
 }
 
 inquirer
   .prompt(questions)
   .then(answers => {
-    fs.unlinkSync('./README.md', err => console.error(err))
+    if (path.join(process.cwd(), './README.md')) fs.unlinkSync('./README.md', err => console.error(err))
     buildReadMe(answers)
   })
   .catch(err => console.error(err))
